+++ src/main/java/org/prgrms/nabimarketbe/NabiMarketBeApplication.java
@@ -4,8 +4,6 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.client.RestTemplate;
 
 @EnableJpaAuditing
@@ -14,12 +12,6 @@ public class NabiMarketBeApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(NabiMarketBeApplication.class, args);
-	}
-
-
-	@Bean
-	public PasswordEncoder passwordEncoder() {
-		return PasswordEncoderFactories.createDelegatingPasswordEncoder();
 	}
 
 	@Bean
