+++ src/main/java/org/prgrms/nabimarketbe/domain/security/jwt/JwtProvider.java
@@ -90,9 +90,9 @@ public class JwtProvider {
         }
     }
 
-    // HTTP Request 의 Header 에서 Token Parsing -> "X-AUTH-TOKEN: jwt"
+    // HTTP Request 의 Header 에서 Token Parsing -> "Authorization: jwt"
     public String resolveToken(HttpServletRequest request) {
-        return request.getHeader("X-AUTH-TOKEN");
+        return request.getHeader("Authorization");
     }
 
     // jwt 의 유효성 및 만료일자 확인
