+++ src/main/java/org/prgrms/nabimarketbe/domain/security/oauth/controller/KOAuthController.java
@@ -3,8 +3,8 @@ package org.prgrms.nabimarketbe.domain.security.oauth.controller;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.prgrms.nabimarketbe.global.model.CommonResult;
-import org.prgrms.nabimarketbe.domain.security.oauth.service.response.ResponseService;
-import org.prgrms.nabimarketbe.domain.user.service.KakaoService;
+import org.prgrms.nabimarketbe.global.ResponseService;
+import org.prgrms.nabimarketbe.domain.security.service.KakaoService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.env.Environment;
 import org.springframework.http.*;
