+++ src/main/java/org/prgrms/nabimarketbe/domain/security/service/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package org.prgrms.nabimarketbe.domain.security.oauth.service;
+package org.prgrms.nabimarketbe.domain.security.service;
 
 import lombok.RequiredArgsConstructor;
 import org.prgrms.nabimarketbe.domain.user.repository.UserJpaRepo;
@@ -7,15 +7,17 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 @RequiredArgsConstructor
 @Service
 public class CustomUserDetailsService implements UserDetailsService {
+
     private final UserJpaRepo userJpaRepo;
 
     @Override
     @Transactional(readOnly = true)
     public UserDetails loadUserByUsername(String userPk) throws UsernameNotFoundException {
         return userJpaRepo.findById(Long.parseLong(userPk))
-                .orElseThrow(() -> new RuntimeException("해당 회원이 없습니다."));
+                .orElseThrow(() -> new RuntimeException("user not found exception"));
     }
 }
