+++ src/main/java/org/prgrms/nabimarketbe/domain/user/controller/SignController.java
@@ -1,22 +1,22 @@
-package org.prgrms.nabimarketbe.domain.sign.controller;
+package org.prgrms.nabimarketbe.domain.user.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.prgrms.nabimarketbe.domain.sign.dto.UserLoginRequestDto;
-import org.prgrms.nabimarketbe.domain.sign.dto.UserSignupRequestDto;
-import org.prgrms.nabimarketbe.domain.sign.dto.UserSocialLoginRequestDto;
-import org.prgrms.nabimarketbe.domain.sign.dto.UserSocialSignupRequestDto;
-import org.prgrms.nabimarketbe.domain.sign.service.SignService;
+
+import org.prgrms.nabimarketbe.domain.security.jwt.dto.TokenRequestDto;
+import org.prgrms.nabimarketbe.domain.user.dto.sign.UserSignupRequestDto;
+import org.prgrms.nabimarketbe.domain.user.dto.sign.UserSocialLoginRequestDto;
+import org.prgrms.nabimarketbe.domain.user.dto.sign.UserSocialSignupRequestDto;
+import org.prgrms.nabimarketbe.domain.user.service.SignService;
 import org.prgrms.nabimarketbe.domain.security.jwt.JwtProvider;
 import org.prgrms.nabimarketbe.domain.security.jwt.dto.TokenDto;
-import org.prgrms.nabimarketbe.domain.security.jwt.dto.TokenRequestDto;
 import org.prgrms.nabimarketbe.domain.security.oauth.dto.social.KakaoProfile;
 import org.prgrms.nabimarketbe.global.model.CommonResult;
 import org.prgrms.nabimarketbe.global.model.SingleResult;
-import org.prgrms.nabimarketbe.domain.security.oauth.service.response.ResponseService;
+import org.prgrms.nabimarketbe.global.ResponseService;
 import org.prgrms.nabimarketbe.domain.user.entity.User;
 import org.prgrms.nabimarketbe.domain.user.repository.UserJpaRepo;
-import org.prgrms.nabimarketbe.domain.user.service.KakaoService;
+import org.prgrms.nabimarketbe.domain.security.service.KakaoService;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -37,25 +37,6 @@ public class SignController {
 
     private final ResponseService responseService;
 
-    @PostMapping("/login")
-    public SingleResult<TokenDto> login(@RequestBody UserLoginRequestDto userLoginRequestDto) {
-        TokenDto tokenDto = signService.login(userLoginRequestDto);
-
-        return responseService.getSingleResult(tokenDto);
-    }
-
-    @PostMapping("/signup")
-    public SingleResult<Long> signup(@RequestBody UserSignupRequestDto userSignupRequestDto) {
-        Long signupId = signService.signup(userSignupRequestDto);
-
-        return responseService.getSingleResult(signupId);
-    }
-
-    @PostMapping("/reissue")
-    public SingleResult<TokenDto> reissue(@RequestBody TokenRequestDto tokenRequestDto) {
-        return responseService.getSingleResult(signService.reissue(tokenRequestDto));
-    }
-
     @PostMapping("/social/login/kakao")
     public SingleResult<TokenDto> loginByKakao(@RequestBody UserSocialLoginRequestDto socialLoginRequestDto) {
         KakaoProfile kakaoProfile = kakaoService.getKakaoProfile(socialLoginRequestDto.getAccessToken());
@@ -70,8 +51,7 @@ public class SignController {
 
     @PostMapping("/social/signup/kakao")
     public CommonResult signupBySocial(@RequestBody UserSocialSignupRequestDto socialSignupRequestDto) {
-        KakaoProfile kakaoProfile =
-                kakaoService.getKakaoProfile(socialSignupRequestDto.getAccessToken());
+        KakaoProfile kakaoProfile = kakaoService.getKakaoProfile(socialSignupRequestDto.getAccessToken());
 
         if (kakaoProfile == null) throw new RuntimeException("해당 회원이 없습니다.");
 
@@ -81,5 +61,12 @@ public class SignController {
                 .build());
 
         return responseService.getSingleResult(userId);
+    }
+
+    //엑세스 토큰 만료시 회원 검증 후 리프레쉬 토큰을 검증해서 액세스 토큰과 리프레시 토큰을 재발급
+    @PostMapping("/reissue")
+    public SingleResult<TokenDto> reissue(
+            @RequestBody TokenRequestDto tokenRequestDto) {
+        return responseService.getSingleResult(signService.reissue(tokenRequestDto));
     }
 }
