+++ src/main/java/org/prgrms/nabimarketbe/domain/user/service/SignService.java
@@ -1,9 +1,8 @@
-package org.prgrms.nabimarketbe.domain.sign.service;
+package org.prgrms.nabimarketbe.domain.user.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.prgrms.nabimarketbe.domain.sign.dto.UserLoginRequestDto;
-import org.prgrms.nabimarketbe.domain.sign.dto.UserSignupRequestDto;
+import org.prgrms.nabimarketbe.domain.user.dto.sign.UserSignupRequestDto;
 import org.prgrms.nabimarketbe.domain.security.jwt.JwtProvider;
 import org.prgrms.nabimarketbe.domain.security.entity.RefreshToken;
 import org.prgrms.nabimarketbe.domain.security.repository.RefreshTokenJpaRepo;
@@ -12,7 +11,6 @@ import org.prgrms.nabimarketbe.domain.security.jwt.dto.TokenRequestDto;
 import org.prgrms.nabimarketbe.domain.user.entity.User;
 import org.prgrms.nabimarketbe.domain.user.repository.UserJpaRepo;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,39 +20,9 @@ import org.springframework.transaction.annotation.Transactional;
 public class SignService {
     private final UserJpaRepo userJpaRepo;
 
-    private final PasswordEncoder passwordEncoder;
-
     private final JwtProvider jwtProvider;
 
     private final RefreshTokenJpaRepo tokenJpaRepo;
-
-    @Transactional
-    public TokenDto login(UserLoginRequestDto userLoginRequestDto) {
-        // 회원 정보 존재하는지 확인
-        User user = userJpaRepo.findByNickname(userLoginRequestDto.getNickname())
-                .orElseThrow(() -> new RuntimeException("해당 회원이 없습니다."));
-
-        // AccessToken, RefreshToken 발급
-        TokenDto tokenDto = jwtProvider.createTokenDto(user.getUserId(), user.getRoles());
-
-        // RefreshToken 저장
-        RefreshToken refreshToken = RefreshToken.builder()
-                .key(user.getUserId())
-                .token(tokenDto.getRefreshToken())
-                .build();
-
-        tokenJpaRepo.save(refreshToken);
-
-        return tokenDto;
-    }
-
-    @Transactional
-    public Long signup(UserSignupRequestDto userSignupDto) {
-        if (userJpaRepo.findByNickname(userSignupDto.getNickname()).isPresent())
-            throw new RuntimeException("이미 존재하는 회원입니다.");
-
-        return userJpaRepo.save(userSignupDto.toEntity()).getUserId();
-    }
 
     @Transactional
     public Long socialSignup(UserSignupRequestDto userSignupRequestDto) {
