+++ src/main/java/org/prgrms/nabimarketbe/global/model/CommonResponse.java
@@ -1,4 +1,4 @@
-package org.prgrms.nabimarketbe.domain.security.oauth.service.response;
+package org.prgrms.nabimarketbe.global.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
