+++ src/test/java/org/prgrms/nabimarketbe/user/service/UserServiceTest.java
@@ -5,8 +5,8 @@ import static org.junit.jupiter.api.Assertions.*;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.prgrms.nabimarketbe.domain.user.service.UserService;
-import org.prgrms.nabimarketbe.domain.sign.dto.UserSignupRequestDto;
-import org.prgrms.nabimarketbe.domain.sign.service.SignService;
+import org.prgrms.nabimarketbe.domain.user.dto.sign.UserSignupRequestDto;
+import org.prgrms.nabimarketbe.domain.user.service.SignService;
 import org.prgrms.nabimarketbe.domain.user.dto.UserRequestDto;
 import org.prgrms.nabimarketbe.domain.user.dto.UserResponseDto;
 import org.prgrms.nabimarketbe.domain.user.entity.User;
